Change `io.spine.time.testing` to `io.spine.testing.time`
